updated method delete

+++ src/routes.js
@@ -97,6 +97,12 @@ export const routes = [
     handler: (req, res) => {
       const { id } = req.params;
 
+      const hasTask = database.selectById('tasks', id);
+
+      if (hasTask === 0) {
+        return res.writeHead(400).end(JSON.stringify({message: 'Task not found'}));
+      }
+
       database.delete('tasks', id);
 
       return res.writeHead(204).end();
